+++ project/src/components/TweetCard.js
@@ -1,4 +1,4 @@
-import React, {useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 import {
   View,
   Text,
@@ -15,21 +15,21 @@ import {
 } from 'react-native';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import Video from 'react-native-video';
-import {createThumbnail} from 'react-native-create-thumbnail';
+import { createThumbnail } from 'react-native-create-thumbnail';
 import DefaultAvatar from '../assets/avatar.png';
 import BottomSheet from './BottomSheet';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import {useNavigation} from '@react-navigation/native';
+import { useNavigation } from '@react-navigation/native';
 import axios from 'axios';
 import config from '../config';
 const serverUrl = config.SERVER_URL;
 
-const TweetCard = ({ tweet, onDeleteSuccess }) => {
+const TweetCard = ({ tweet, onDeleteSuccess, comments }) => {
   const [liked, setLiked] = useState(tweet.isLiked);
-  const [likesCount, setLikesCount] = useState(tweet.likesCount);
-  const [bookmarked, setBookmarked] = useState(tweet.isBookmarked);
-  const [bookMarksCount, setBookMarksCount] = useState(tweet.bookMarksCount);
-  const [commentsCount] = useState(tweet.commentsCount);
+  const [likesCount, setLikesCount] = useState(tweet.likesCount || 0);
+  const [bookmarked, setBookmarked] = useState(tweet.isBookmarked || false);
+  const [bookMarksCount, setBookMarksCount] = useState(tweet.bookMarksCount || 0);
+  const [commentsCount] = useState(tweet.commentsCount || 0);
   const [isModalVisible, setIsModalVisible] = useState(false);
   const [modalMediaUri, setModalMediaUri] = useState('');
   const [thumbnails, setThumbnails] = useState({});
@@ -54,7 +54,7 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
       for (const media of tweet.media || []) {
         if (media.type === 'video' && media.uri) {
           try {
-            const {path} = await createThumbnail({url: media.uri});
+            const { path } = await createThumbnail({ url: media.uri });
             newThumbnails[media.uri] = path;
           } catch (error) {
             console.log('Error generating thumbnail:', error);
@@ -174,7 +174,15 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
   };
 
   const handleCommentPress = () => {
-    navigator.navigate('Comment', { postId: tweet.id, idUser: tweet.idUser, profilePicture: tweet.profilePicture, emailUser: tweet.emailUser });
+    navigator.navigate('ViewPost', {
+      tweet,
+      postId: tweet.id,
+      idUser: tweet.idUser,
+      profilePicture: tweet.profilePicture,
+      emailUser: tweet.emailUser,
+      comments,
+      focusCommentInput: true,
+    });
   };
 
   const openMediaPreview = uri => {
@@ -239,7 +247,7 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
     return `${date.getDate()}-${date.getMonth() + 1}-${date.getFullYear()}`;
   };
 
-  const renderMediaItem = ({item}) => {
+  const renderMediaItem = ({ item }) => {
     if (!item.uri) {
       return null;
     }
@@ -250,8 +258,8 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
           ? styles.singleMediaVideo
           : styles.singleMediaImage
         : item.type === 'video'
-        ? styles.tweetVideo
-        : styles.tweetImage;
+          ? styles.tweetVideo
+          : styles.tweetImage;
 
     return (
       <TouchableOpacity
@@ -259,7 +267,7 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
         style={styles.mediaContainer}>
         {item.type === 'image' ? (
           <Image
-            source={{uri: item.uri}}
+            source={{ uri: item.uri }}
             style={mediaStyle}
             onError={() => console.log('Failed to load image')}
           />
@@ -267,7 +275,7 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
           <TouchableOpacity
             onPress={() => openMediaPreview(item.uri)}
             style={styles.videoContainer}>
-            <Image source={{uri: thumbnails[item.uri]}} style={mediaStyle} />
+            <Image source={{ uri: thumbnails[item.uri] }} style={mediaStyle} />
             <MaterialCommunityIcons
               name="play-circle-outline"
               size={40}
@@ -307,7 +315,7 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
       <View style={styles.userInfo}>
         <TouchableOpacity onPress={handleProfilePress}>
           <Image
-            source={tweet.userAvatar ? {uri: tweet.userAvatar} : DefaultAvatar}
+            source={tweet.userAvatar ? { uri: tweet.userAvatar } : DefaultAvatar}
             style={styles.avatar}
           />
         </TouchableOpacity>
@@ -346,7 +354,7 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
             <BottomSheet
               onClose={() => {
                 setShowBottomSheet(false);
-                onDeleteSuccess(); // Panggil onDeleteSuccess setelah BottomSheet tertutup
+                onDeleteSuccess();
               }}
               username={tweet.userHandle}
               postId={tweet.id}
@@ -417,15 +425,15 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
             <View style={styles.modalContainer}>
               {modalMediaUri ? (
                 modalMediaUri.endsWith('.jpg') ||
-                modalMediaUri.endsWith('.png') ? (
+                  modalMediaUri.endsWith('.png') ? (
                   <Image
-                    source={{uri: modalMediaUri}}
+                    source={{ uri: modalMediaUri }}
                     style={styles.modalImage}
                     onError={() => console.log('Failed to load image')}
                   />
                 ) : (
                   <Video
-                    source={{uri: modalMediaUri}}
+                    source={{ uri: modalMediaUri }}
                     style={styles.modalImage}
                     controls
                     resizeMode="contain"
@@ -440,7 +448,7 @@ const TweetCard = ({ tweet, onDeleteSuccess }) => {
   );
 };
 
-const InteractionButton = ({icon, color, count, onPress}) => (
+const InteractionButton = ({ icon, color, count, onPress }) => (
   <TouchableOpacity style={styles.actionButton} onPress={onPress}>
     <MaterialCommunityIcons name={icon} size={20} color={color} />
     <Text style={styles.actionText}>{count}</Text>
@@ -452,7 +460,7 @@ const styles = StyleSheet.create({
     backgroundColor: '#fff',
     padding: 10,
     shadowColor: '#000',
-    shadowOffset: {width: 0, height: 2},
+    shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.1,
     shadowRadius: 4,
     elevation: 1,
